Moved server to separate module
