+++ go.mod
@@ -1,0 +1,3 @@
+module cmd/main
+
+go 1.16
